pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "Agentic job search assistant that finds relevant Danish job listings"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "anthropic>=0.40.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "resend>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
job-agent = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
